# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LabelLens API",
    "version": "1.0.0",
    "description": "Contract for the LabelLens v1 HTTP API. The current scope is food search, barcode lookup, menu calculation, saved menus and favorites. Compare, recipes, pantry inventory and export/PDF are not part of this contract."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local gateway/API during development"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/session": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "startSession",
        "summary": "Start an authenticated session by logging in or registering.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionSuccessResponse"
                }
              }
            }
          },
          "202": {
            "description": "Registration created and waiting for confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfirmationRequiredResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid session request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "confirmSession",
        "summary": "Confirm a Cognito registration code and finish sign-in.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthConfirmationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account confirmed and session started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid confirmation request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "requestPasswordReset",
        "summary": "Send a Cognito password recovery code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPasswordResetStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recovery code sent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPasswordResetStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid password reset request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Password reset is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "confirmPasswordReset",
        "summary": "Confirm a Cognito password recovery code and set a new password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPasswordResetConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPasswordResetConfirmResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid password update request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Password update is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "getCurrentUser",
        "summary": "Return the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/AuthUser"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "dev-token-local-now-cognito-later"
      }
    },
    "schemas": {
...
      "AuthUser": {
        "type": "object",
        "required": [
          "userId",
          "displayName"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "AuthSessionRequest": {
        "type": "object",
        "required": [
          "mode",
          "email",
          "password"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "login",
              "register"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 120
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          }
        },
        "additionalProperties": false
      },
      "AuthConfirmationRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "confirmationCode"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 120
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "confirmationCode": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32
          }
        },
        "additionalProperties": false
      },
      "AuthPasswordResetStartRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 120
          }
        },
        "additionalProperties": false
      },
      "AuthPasswordResetStartResponse": {
        "type": "object",
        "required": [
          "nextStep",
          "email",
          "message"
        ],
        "properties": {
          "nextStep": {
            "type": "string",
            "enum": [
              "reset-password"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "deliveryDestination": {
            "type": "string"
          },
          "deliveryMedium": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AuthPasswordResetConfirmRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "confirmationCode"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 120
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "confirmationCode": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32
          }
        },
        "additionalProperties": false
      },
      "AuthPasswordResetConfirmResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "AuthSessionSuccessResponse": {
        "type": "object",
        "required": [
          "tokenType",
          "accessToken",
          "user",
          "authMode"
        ],
        "properties": {
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "accessToken": {
            "type": "string"
          },
          "authMode": {
            "type": "string",
            "enum": [
              "demo",
              "cognito"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
      "AuthConfirmationRequiredResponse": {
        "type": "object",
        "required": [
          "nextStep",
          "email",
          "message"
        ],
        "properties": {
          "nextStep": {
            "type": "string",
            "enum": [
              "confirm"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "deliveryDestination": {
            "type": "string"
          },
          "deliveryMedium": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "correlationId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "details": {}
        }
      }
    }
  }
}